+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ManualHandlerMapping.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.mapping;
+package com.interface21.webmvc.servlet.mvc.asis;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.mapping.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
